ci(github): fix workflow syntax and drive Windows job through dev.ps1
The workflow file was rejected outright: a stray escape had corrupted the
inline MSVC setup script. The Windows job now calls scripts/dev.ps1, which
already enters the MSVC environment, so CI exercises the same entry point a
developer uses and a break in that script fails here rather than locally.

Also silences three clang-tidy checks that fire constantly on integer pixel
math without finding anything, and fixes the widening casts it did find.

+++ src/renderer/framebuffer.hpp
@@ -19,7 +19,8 @@ struct Sprite {
     [[nodiscard]] constexpr int bytesPerRow() const noexcept { return (width + 7) / 8; }
     [[nodiscard]] constexpr bool valid() const noexcept {
         return width > 0 && height > 0 &&
-               bits.size() >= static_cast<std::size_t>(bytesPerRow() * height);
+               bits.size() >=
+                   static_cast<std::size_t>(bytesPerRow()) * static_cast<std::size_t>(height);
     }
 };
 
@@ -73,4 +74,4 @@ private:
     std::array<std::uint8_t, kByteCount> pixels_{};
 };
 
-}  // namespace wumpo::renderer
+} // namespace wumpo::renderer
